CEAC-serverless/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
flatbuffers==2.0
idna==3.3
numpy==1.21.4
onnxruntime==1.9.0
Pillow==8.4.0
protobuf==3.19.1
requests==2.26.0
soupsieve==2.3
urllib3==1.26.7
